Changes:  - Intel 11 C++ compiler (on Linux) can be used for regression testing
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5444 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/squadt/libraries/tipi/include/tipi/basic_datatype.hpp
@@ -331,7 +331,7 @@ namespace tipi {
          * \return *this
          **/
         inline enumeration< C >& add(const size_t v, std::string const& s) {
-          get_single_instance().add(static_cast< const size_t > (v), s);
+          get_single_instance().add(v, s);
 
           return *this;
         }
